feat(app): create labware card and labware landing
fix #8861
