+++ labware-library/src/filters.tsx
@@ -5,7 +5,7 @@ import flatMap from 'lodash/flatMap'
 import pickBy from 'lodash/pickBy'
 import uniq from 'lodash/uniq'
 
-import { getAllDefinitions } from './definitions'
+import { getAllDefinitions } from '@opentrons/shared-data'
 import { getPublicPath } from './public-path'
 
 import type { Location } from 'history'
